requirements.txt
Flask~=1.1.2
omxplayer-wrapper~=0.3.3
youtube_dl>=2020.7.28
